+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.44.0](https://github.com/EOCOnline/rangertrak/commits/main) (2026-08-25)
+
+### Features
+
+* **help,shared,field-reports,rangers:** E-84 - Help page tabs, on-page text fixes, deduplicated grid-keyboard-help component, FIELD-GUIDE.md and README.md rewrite per the documentation audit
+
+### Bug Fixes
+
+* **mapping,mapLeaflet,map,entry:** map marker tooltips show short local time + "N min ago", not the raw `Date` object
+
+## [0.43.1](https://github.com/EOCOnline/rangertrak/commit/4c8d3dbe1be66bd9de2852626dc20dfd070a0467) (2026-08-24)
+
+### Bug Fixes
+
+* **mapLeaflet,mapping:** sequential-name hash collision, offline controls out of the map, minor UI polish ([4c8d3db](https://github.com/EOCOnline/rangertrak/commit/4c8d3dbe1be66bd9de2852626dc20dfd070a0467))
+
+## [0.43.0](https://github.com/EOCOnline/rangertrak/commit/db9f3dd76ede29c6a8152550fc4d1a1f1440877d) (2026-08-24)
+
+### Features
+
+* **entry,header,services,e2e:** E-83 - reopenable Entry welcome panel ([db9f3dd](https://github.com/EOCOnline/rangertrak/commit/db9f3dd76ede29c6a8152550fc4d1a1f1440877d))
+
 ## [0.42.0](https://github.com/EOCOnline/rangertrak/commit/64363d446e33dbf75c1cbbdfd52ec0cc79bf5c2f) (2026-08-24)
 
 ### Features